# Decrease doc annotation time from 0.5 sec to 0.005 sec