--- a/tools/lsp-wake/astree.h
+++ b/tools/lsp-wake/astree.h
@@ -63,11 +63,14 @@ class ASTree {
   struct SymbolUsage {
     Location usage;
     Location definition;
-    SymbolUsage(Location _usage, Location _definition);
+    SymbolUsage(Location _usage, Location _definition)
+        : usage(std::move(_usage)), definition(std::move(_definition)) {}
   };
 
   struct Comment {
-    Comment(std::string _comment_text, Location _location, int level);
+    Comment() : location("") {}
+    Comment(std::string _comment_text, Location _location, int _level)
+        : comment_text(std::move(_comment_text)), location(std::move(_location)), level(_level) {}
 
     std::string comment_text;
     Location location;
@@ -78,7 +81,7 @@ class ASTree {
   std::vector<SymbolDefinition> definitions;
   std::vector<SymbolUsage> usages;
   std::vector<SymbolDefinition> packages;
-  std::vector<Comment> comments;
+  std::unordered_map<std::string, Comment> comments;
 
   class LSPReporter : public DiagnosticReporter {
    private:
@@ -102,11 +105,11 @@ class ASTree {
 
   static std::string sanitizeComment(std::string comment);
 
-  static std::string composeOuterComment(std::vector<std::pair<std::string, int>> comment);
+  std::string sanitizeMultilineComment(const std::vector<Comment> &comment);
 
   static void emplaceComment(std::vector<std::pair<std::string, int>> &comment,
                              const std::string &text, int level);
 
-  void recordSameLocationDefinition(std::vector<SymbolDefinition>::iterator &definitions_iterator);
+  void mergeSameLineDefinitions(const SymbolDefinition &merge_from, SymbolDefinition *into);
 };
 #endif