--- a/tools/lsp-wake/astree.h
+++ b/tools/lsp-wake/astree.h
@@ -1,3 +1,5 @@
+
+
 /* Wake Language Server Protocol implementation
  *
  * Copyright 2021 SiFive, Inc.