--- a/tools/lsp-wake/astree.h
+++ b/tools/lsp-wake/astree.h
@@ -1,5 +1,3 @@
-
-
 /* Wake Language Server Protocol implementation
  *
  * Copyright 2021 SiFive, Inc.