--- a/tools/lsp-wake/json_converter.cpp
+++ b/tools/lsp-wake/json_converter.cpp
@@ -318,7 +318,6 @@ JAST hoverInfoToJSON(JAST receivedMessage, const std::vector<SymbolDefinition> &
         value += "**" + introduced.first + ": " + introduced.second + "**\n\n";
       }
     }
-    value += def.outerDocumentation + "\n\n";
   }
 
   if (value.empty()) {