… tools/lsp-wake/astree.cpp