This PR optimizes the documentation discovery of the LSP core loop. The core loop is what gives the most up to date info to the user and is expected to be snappy. Wake's LSP's core loop is painfully slow at ~2 seconds. Previously the documentation step took 0.5 seconds, most of which was spent in a O(n^2) deletion step. The step is now floating around 0.008  seconds. 

Old profile on my fast disk machine

```
find files: 0.126626 seconds
front: 0.000044 seconds
cst/dst loop: 0.242263 seconds
flatten exports: 0.001422 seconds
packages loop: 0.000264 seconds
bind refs: 0.546380 seconds
callback: 0.000275 seconds
explore: 0.114686 seconds
sort: 0.008359 seconds
fillDocumentation: 0.606422 seconds
Refreshed project in 1.824396 seconds (due to initialized)
```

New profile on my fast disk machine
```
find files: 0.125346 seconds
front: 0.000042 seconds
cst/dst loop: 0.205993 seconds
flatten exports: 0.001369 seconds
packages loop: 0.000225 seconds
bind refs: 0.563156 seconds
callback: 0.000269 seconds
explore: 0.114466 seconds
sort: 0.000000 seconds
fillDocumentation: 0.009003 seconds
Refreshed project in 1.195170 seconds (due to initialized)
```